+++ src/main/java/tools/IntervalBuilder.java
@@ -63,7 +63,7 @@ public class IntervalBuilder {
          newLimitX2 = newTo;
       }
 
-      intervalLimits.add(st, new Line2D.Float((float) newLimitX1, 0, (float) newLimitX2, 0));
+      intervalLimits.add(st, new Line2D.Float(newLimitX1, 0, newLimitX2, 0));
    }
 
    private int bsearchInterval(int point) {
